pyetrade-main/requirements.txt
requests
requests_oauthlib
xmltodict
jxmlease
flask
azure-identity
azure-keyvault-secrets
selenium
chromium-chromedriver